// src/shared/dataprovider/api/model/balanceTransferDataDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BalanceTransferBreakdownDestinationDTO } from './balanceTransferBreakdownDestinationDTO';
import { BalanceTransferBreakdownSourceDTO } from './balanceTransferBreakdownSourceDTO';
import { TopUpStatusEnum } from './topUpStatusEnum';
import { TransactionTypeEnum } from './transactionTypeEnum';

export interface BalanceTransferDataDTO { 
    Id?: number;
    ClientId?: number;
    ClientName?: string;
    SourceChannelId?: number;
    SourceChannelName?: string;
    DestinationChannelId?: number;
    DestinationChannelName?: string;
    TransactionNumber?: string;
    Requestee?: string;
    RequesteeName?: string;
    Status?: TopUpStatusEnum;
    StatusDescription?: string;
    Amount?: number;
    CreatedDate?: Date;
    CompletedDate?: Date;
    ApproverId?: number;
    ApproverName?: string;
    Remarks?: string;
    TransferType?: number;
    TransferTypeDescription?: string;
    SourceMerchants?: Array<BalanceTransferBreakdownSourceDTO>;
    DestinationMerchants?: Array<BalanceTransferBreakdownDestinationDTO>;
}


export  interface BTparams {
    ClientId? :number;
    StartDate?:Date;
    EndDate? : Date;
    PageSize? : number;
    PageNumber? : number;
    SearchKeyword? : string;
    Statuses? : string;
    TransferTypes: string;
  }